requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
coverage==7.6.4
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.3
django-cache-url==3.4.5
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
environs==11.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.23.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
